smart_vehicle_backend/requirements.txt
flask
flask-login
flask-sqlalchemy
flask_cors
requests
opencv-python
numpy
python-dotenv
mediapipe